+++ src/Physics/Pre_calculation/Pre_Calculation_Factory.jl
@@ -17,7 +17,7 @@ using .Shape_Tensor
 export compute
 export init_pre_calculation
 
-function compute(datamanager::Module, nodes::Union{SubArray,Vector{Int64}}, options, time::Float32, dt::Float32)
+function compute(datamanager::Module, nodes::Union{SubArray,Vector{Int64}}, options, time::Float64, dt::Float64)
 
     if options["Deformed Bond Geometry"]
         datamanager = Bond_Deformation.compute(datamanager, nodes, time)
@@ -40,21 +40,21 @@ end
 function init_pre_calculation(datamanager, options)
     dof = datamanager.get_dof()
     if options["Deformed Bond Geometry"]
-        bond_defN, bond_defNP1 = datamanager.create_bond_field("Deformed Bond Geometry", Float32, dof + 1)
+        bond_defN, bond_defNP1 = datamanager.create_bond_field("Deformed Bond Geometry", Float64, dof + 1)
     end
     if options["Shape Tensor"]
-        shapeTensor = datamanager.create_constant_node_field("Shape Tensor", Float32, "Matrix", dof)
-        invShapeTensor = datamanager.create_constant_node_field("Inverse Shape Tensor", Float32, "Matrix", dof)
+        shapeTensor = datamanager.create_constant_node_field("Shape Tensor", Float64, "Matrix", dof)
+        invShapeTensor = datamanager.create_constant_node_field("Inverse Shape Tensor", Float64, "Matrix", dof)
     end
     if options["Deformation Gradient"]
-        defGrad = datamanager.create_constant_node_field("Deformation Gradient", Float32, "Matrix", dof)
+        defGrad = datamanager.create_constant_node_field("Deformation Gradient", Float64, "Matrix", dof)
     end
     if options["Bond Associated Shape Tensor"]
-        bondShapeTensor = datamanager.create_constant_bond_field("Bond Associated Shape Tensor", Float32, "Matrix", dof)
-        invBondShapeTensor = datamanager.create_constant_bond_field("Inverse Bond Associated Shape Tensor", Float32, "Matrix", dof)
+        bondShapeTensor = datamanager.create_constant_bond_field("Bond Associated Shape Tensor", Float64, "Matrix", dof)
+        invBondShapeTensor = datamanager.create_constant_bond_field("Inverse Bond Associated Shape Tensor", Float64, "Matrix", dof)
     end
     if options["Bond Associated Deformation Gradient"]
-        bondDefGrad = datamanager.create_constant_bond_field("Bond Associated Deformation Gradient", Float32, "Matrix", dof)
+        bondDefGrad = datamanager.create_constant_bond_field("Bond Associated Deformation Gradient", Float64, "Matrix", dof)
     end
     return datamanager
 end
