+++ src/Physics/Pre_calculation/Pre_calculation_template/pre_calculation_template.jl
@@ -33,8 +33,8 @@ end
         - `datamanager::Data_manager`: Datamanager.
         - `nodes::Union{SubArray,Vector{Int64}}`: List of block nodes.
         - `Pre_calculation_parameter::Dict(String, Any)`: Dictionary with material parameter.
-        - `time::Float32`: The current time.
-        - `dt::Float32`: The current time step.
+        - `time::Float64`: The current time.
+        - `dt::Float64`: The current time step.
    Returns:
         - - `datamanager::Data_manager`: Datamanager.
    Example:
@@ -42,7 +42,7 @@ end
      ```
    """
 
-function pre_calculation(datamanager::Module, nodes::Union{SubArray,Vector{Int64}}, Pre_calculation_parameter::Dict, time::Float32, dt::Float32)
+function pre_calculation(datamanager::Module, nodes::Union{SubArray,Vector{Int64}}, Pre_calculation_parameter::Dict, time::Float64, dt::Float64)
   @info "Please write a possible precalculation routines in pre_calculation_name()."
   @info "You can call your routine within the yaml file."
   @info "Fill the compute_force(datamanager, nodes, Pre_calculation_parameter, time, dt) function."
